+++ qunit/main_phone.js
@@ -1,8 +1,8 @@
 define([
 	"qunit",
-	"tests_phone_world",
-	"tests_phonebe",
-	"tests_phonenl",
+	// "tests_phone_world",
+	// "tests_phonebe",
+	// "tests_phonenl",
 	"tests_phoneru"
 ], function(qunit) {
 	qunit.load();